PageImageCandidate: Cast handlerWidth to int
Bug: T312522
(cherry picked from commit 7e8167f3be62e1928e257546d64aebb55b514249)

+++ includes/PageImageCandidate.php
@@ -41,7 +41,7 @@ class PageImageCandidate implements JsonSerializable {
 		$instance->fullWidth = $file->getWidth() ?? 0;
 		$instance->fullHeight = $file->getHeight() ?? 0;
 		if ( isset( $fileParams['handler']['width'] ) ) {
-			$instance->handlerWidth = $fileParams['handler']['width'] ?? 0;
+			$instance->handlerWidth = (int)( $fileParams['handler']['width'] ?? 0 );
 		}
 		return $instance;
 	}
